# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Title",
    "version": "1.0.0",
    "description": "Your API description"
  },
  "paths": {
...
    "/transactions": {
      "get": {
        "summary": "Retrieve all transactions",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "200": {
            "description": "A list of transactions"
          }
        }
      },
      "post": {
        "summary": "Create a new transaction",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "ID of the order for the transaction"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Amount of the transaction"
                  }
                },
                "example": {
                  "orderId": 1,
                  "amount": 100.5
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created transaction object"
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "summary": "Retrieve a single transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the transaction to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "A transaction object"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "put": {
        "summary": "Update a transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the transaction to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "Updated ID of the order for the transaction"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Updated amount of the transaction"
                  }
                },
                "example": {
                  "orderId": "1 (Updated)",
                  "amount": 120.75
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated transaction object"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the transaction to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Transaction deleted successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        }
      }
    },
...
  },
...
}